Experimental support for GitHub token authentication
This comes from the desire to be more fine-grained in the permissions
given to the github management automation.

E.g. With GitHub Apps, there's no way to give both the "manage teams"
permission without also the "remove org owners" permission. Similarly,
the "manage repos" permission also comes with the "remove repos"
permission. only what's required.

In comparison, with a personal access token of a user, you can give that
user specific access levels on the repos it needs access to, or only add
it to specific teams it needs to manage.

For now, personal access tokens are limited to team member management,
because unless you make the user an org admin, you run into problems for
pretty much any other functionality.